Update hinted-mods.py
Now partially reversing the previous declaration of all import files also as "implicit imports":
Only modules starting with ``matplotlib.backends`` are declared as implicit imports as well.

Now all matplotlib  imports are ignored if there not also exists an import for ``matplotlib.backends``. This is because some packages do a matplotlib import just to see if it is installed, but can live without it, if the main application does not use a matplotlib backend.